Move info links handling and version handling from get_yaml_data to freee_a11y_gl
