+++ tools/yaml2x/libs/freee_a11y_gl/freee_a11y_gl/__init__.py
@@ -2,3 +2,5 @@ from .classes import *
 from .constants import *
 from .source import get_src_path
 from .initializer import setup_instances
+from .info_utils import get_info_links
+from .version_utils import get_version_info
