+++ tools/yaml2x/libs/freee_a11y_gl/freee_a11y_gl/info_utils.py
@@ -1,8 +1,5 @@
 """
-Utility functions for YAML to JSON conversion.
-
-This module provides utility functions for handling version information
-and info links processing from pickle files.
+Utility functions for info and link processing.
 """
 
 import pickle
@@ -13,8 +10,8 @@ from typing import Dict, Any, List
 LANGUAGES: List[str] = ['ja', 'en']
 PICKLE_PATH: str = 'build/doctrees/environment.pickle'
 
-class UtilsError(Exception):
-    """Custom exception for utility-related errors."""
+class InfoUtilsError(Exception):
+    """Custom exception for info utilities related errors."""
     pass
 
 def get_info_links(basedir: str, baseurl: str = '') -> Dict[str, Any]:
@@ -29,7 +26,7 @@ def get_info_links(basedir: str, baseurl: str = '') -> Dict[str, Any]:
         Dictionary containing extracted labels and their associated information
         
     Raises:
-        UtilsError: If pickle file cannot be loaded or processed
+        InfoUtilsError: If pickle file cannot be loaded or processed
     """
     info: Dict[str, Dict[str, Dict[str, str]]] = {}
     path_prefix = {
@@ -43,7 +40,7 @@ def get_info_links(basedir: str, baseurl: str = '') -> Dict[str, Any]:
             with open(pickle_path, 'rb') as f:
                 doctree = pickle.load(f)
         except Exception as e:
-            raise UtilsError(f'Failed to load pickle file {pickle_path}: {str(e)}')
+            raise InfoUtilsError(f'Failed to load pickle file {pickle_path}: {str(e)}')
             
         labels = doctree.domaindata['std']['labels']
         for label in labels:
@@ -59,30 +56,3 @@ def get_info_links(basedir: str, baseurl: str = '') -> Dict[str, Any]:
             info[label]['url'][lang] = f'{baseurl}/{path_prefix[lang]}{labels[label][0]}.html#{labels[label][1]}'
 
     return info
-
-def get_version_info(basedir: str) -> Dict[str, str]:
-    """
-    Extract version information from version.py file.
-    
-    Args:
-        basedir: Base directory containing version.py
-        
-    Returns:
-        Dictionary containing version information
-        
-    Raises:
-        UtilsError: If version file cannot be read or executed
-    """
-    try:
-        version_data: Dict[str, str] = {}
-        version_file = Path(basedir) / 'version.py'
-        
-        if not version_file.is_file():
-            raise UtilsError(f"Version file not found: {version_file}")
-            
-        with open(version_file, encoding='utf-8') as f:
-            exec(f.read(), version_data)
-            
-        return version_data
-    except Exception as e:
-        raise UtilsError(f"Failed to read version information: {str(e)}")
